Update "uploadPerform" to set "Initial revision" if it applies
This commit ensures that the first revision of new item will always be
associated with the comment "Initial revision". This applies to new files
uploaded using either the web apis or the web interface.

The behavior is now consistent with the initial message set in
UploadComponent::createLinkItem() and ApisystemComponent::uploadGeneratetoken()